add board list

+++ src/main/java/necrojan/games/hangman/model/Board.java
@@ -1,25 +1,21 @@
 package necrojan.games.hangman.model;
 
 import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 import lombok.Value;
 import lombok.With;
-import lombok.extern.slf4j.Slf4j;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
 import java.util.List;
 
 @Value
 @AllArgsConstructor
 public class Board {
     Long userId;
-    int total;
+    Long total;
 
     @With
     List<String> badges;
 
-    public Board(Long userId, int total) {
+    public Board(Long userId, Long total) {
         this.userId = userId;
         this.total = total;
         this.badges = List.of();
